requirements.txt
absl-py==2.1.0
accelerate==1.5.2
addict==2.4.0
alembic==1.14.0
annotated-types==0.7.0
ansi2html==1.9.2
attrs==23.2.0
basicsr==1.4.2
bcrypt==4.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cmaes==0.11.1
coloredlogs==15.0.1
colorlog==6.9.0
conda-pack==0.8.0
contourpy==1.3.1
cryptography==43.0.3
cycler==0.12.1
dash==2.12.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==4.4.2
deepspeed==0.16.4
einops==0.8.1
exceptiongroup==1.2.2
filelock==3.18.0
Flask==2.2.5
flatbuffers==25.2.10
fonttools==4.55.0
fsspec==2025.3.0
ftfy==6.3.1
future==1.0.0
greenlet==3.1.1
grpcio==1.73.1
hjson==3.1.0
huggingface-hub==0.29.3
humanfriendly==10.0
idna==3.10
imageio==2.37.0
iniconfig==2.0.0
invoke==1.7.3
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.0
jsonschema==4.19.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
lazy_loader==0.4
lmdb==1.7.2
lxml==5.2.1
Mako==1.1.0
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
mmcv==1.7.1
mmengine==0.10.7
mock==3.0.5
mpmath==1.3.0
msgpack==1.1.0
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.3
numpy==1.26.4
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-ml-py==12.570.86
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
onnx==1.17.0
onnxruntime==1.21.0
open_clip_torch==2.32.0
opencv-python==4.5.4.58
optuna==3.3.0
packaging==24.2
pandas==2.0.1
paramiko==3.4.0
pathlib2==2.3.6
peft==0.7.1
pillow==11.3.0
platformdirs==4.3.8
plotly==5.24.1
pluggy==1.5.0
protobuf==5.28.3
psutil==5.6.4
py-cpuinfo==9.0.0
pyaml==24.9.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.2
PyNaCl==1.5.0
pyparsing==3.2.0
pytest==8.1.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==5.3
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
retrying==1.3.4
rich==14.0.0
rpds-py==0.21.0
safetensors==0.4.3
scikit-image==0.25.2
scikit-learn==1.5.2
scikit-optimize==0.9.0
scipy==1.15.3
six==1.16.0
SQLAlchemy==2.0.36
sympy==1.14.0
tabulate==0.9.0
tb-nightly==2.21.0a20250723
tenacity==9.0.0
tensorboard-data-server==0.7.2
termcolor==3.1.0
tflite==2.3.0
threadpoolctl==3.5.0
tifffile==2025.5.10
tiktoken==0.9.0
timm==0.6.13
tokenizers==0.15.2
tomli==2.1.0
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
transformers==4.37.2
transformers-stream-generator==0.0.4
triton==3.3.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==2.2.3
XlsxWriter==1.2.2
yapf==0.43.0
